email_enrichment/requirements.txt
dnspython>=2.4
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
